Testing reveals...
I need to include a target course for sequences and the like. That is,
it is not enough to ask "do you have this sequence?" It is "do you have
this sequence *for this particular course*."

I had this originally, but I dropped it as I did the rewrite. Factoring
it back in.